requirements.txt
anthropic
python-dotenv
flask
flask-cors